Rename classes.
