+++ src/main/java/info/coremodding/craftenchants/item/recipes/EnchantsArmor.java
@@ -1,39 +1,40 @@
 /**
- * Copyright 2014 by CoreModding under GNU Lesser General Public License (LGPLv3)
- * http://www.gnu.org/licenses/gpl.html
- * http://www.gnu.org/licenses/lgpl.html
- */
+  * Copyright 2014 by CoreModding under GNU Lesser General Public License (LGPLv3)
+  * http://www.gnu.org/licenses/gpl.html
+  * http://www.gnu.org/licenses/lgpl.html
+  */
 package info.coremodding.craftenchants.item.recipes;
 
 import info.coremodding.craftenchants.item.ItemCE;
 import info.coremodding.craftenchants.item.enchants.ReinforcedNock;
 
+import net.minecraft.enchantment.Enchantment;
 import net.minecraft.inventory.InventoryCrafting;
 import net.minecraft.item.Item;
-import net.minecraft.item.ItemBow;
+import net.minecraft.item.ItemArmor;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.IRecipe;
 import net.minecraft.world.World;
 
-public class EnchantBow implements IRecipe {
+public class EnchantsArmor implements IRecipe {
 	private ItemCE enchantingItem;
-	private ItemStack enchantableBow;
-
+	private ItemStack enchantableArmor;
+	
 	/**
 	 * @see net.minecraft.item.crafting.IRecipe#matches(net.minecraft.inventory.InventoryCrafting, net.minecraft.world.World)
 	 */
 	@Override
 	public boolean matches(InventoryCrafting craftingMatrix, World craftWorld) {
-		boolean containsBow = false;
+		boolean containsArmor = false;
 		boolean containsEnchant = false;
 		ItemStack slottedItemStack;
 		for (int slot = 0; slot < craftingMatrix.getSizeInventory(); slot++) {
 			slottedItemStack = craftingMatrix.getStackInSlot(slot);
 			if (slottedItemStack != null) {
 				Item slottedItem = slottedItemStack.getItem();
-				if (slottedItem instanceof ItemBow) {
-					enchantableBow = slottedItemStack.copy();
-					containsBow = true;
+				if (slottedItem instanceof ItemArmor) {
+					enchantableArmor = slottedItemStack.copy();
+					containsArmor = true;
 				}
 				if (slottedItem instanceof ReinforcedNock) {
 					enchantingItem = (ItemCE)slottedItem;
@@ -41,7 +42,8 @@ public class EnchantBow implements IRecipe {
 				}
 			}
 		}
-		return containsEnchant &&  containsBow;
+		
+		return containsEnchant && containsArmor;
 	}
 
 	/**
@@ -49,16 +51,31 @@ public class EnchantBow implements IRecipe {
 	 */
 	@Override
 	public ItemStack getCraftingResult(InventoryCrafting inventorycrafting) {
+		return applyEnchantment();
+	}
+
+	private ItemStack applyEnchantment() {
 		return shouldEnchant() ? applyEnchant() : null;
 	}
-
+	
 	private boolean shouldEnchant() {
-		return !enchantableBow.isItemEnchanted();
+		boolean willEnchant = false;
+		if (((ItemArmor)(enchantableArmor.getItem())).armorType == 0) {
+			if (enchantingItem.getEnchantType().equals(Enchantment.respiration)) willEnchant = true;
+		}
+		if (((ItemArmor)(enchantableArmor.getItem())).armorType == 1) {
+			if (enchantingItem.getEnchantType().equals(Enchantment.blastProtection)) willEnchant = true;
+			if (enchantingItem.getEnchantType().equals(Enchantment.thorns)) willEnchant = true;
+		}
+		if (((ItemArmor)(enchantableArmor.getItem())).armorType == 3) {
+			if (enchantingItem.getEnchantType().equals(Enchantment.featherFalling)) willEnchant = true;
+		}
+		return !enchantableArmor.isItemEnchanted() && willEnchant;
 	}
-
+	
 	private ItemStack applyEnchant() {
-		enchantableBow.addEnchantment(enchantingItem.getEnchantType(), enchantingItem.getEnchantLevel());
-		return enchantableBow;
+		enchantableArmor.addEnchantment(enchantingItem.getEnchantType(), enchantingItem.getEnchantLevel());
+		return enchantableArmor;
 	}
 
 	/**
